+++ benchmark/decoding_probability/codes.hpp
@@ -8,7 +8,7 @@
 
 #include <kodo/rlnc/full_vector_codes.hpp>
 #include <kodo/linear_block_decoder_delayed.hpp>
-
+#include <kodo/sparse_uniform_generator.hpp>
 
 namespace kodo
 {
@@ -105,6 +105,40 @@ namespace kodo
                    > > > > > > > > > > > > > > > >
     { };
 
+    /// RLNC encoder using a density based random generator, which can be
+    /// used to control the density i.e. the number of non-zero elements in
+    /// the encoding vector.
+    template<class Field>
+    class sparse_full_rlnc_encoder :
+        public // Payload Codec API
+               payload_encoder<
+               // Codec Header API
+               systematic_encoder<
+               symbol_id_encoder<
+               // Symbol ID API
+               plain_symbol_id_writer<
+               // Coefficient Generator API
+               sparse_uniform_generator<
+               // Codec API
+               encode_symbol_tracker<
+               zero_symbol_encoder<
+               linear_block_encoder<
+               storage_aware_encoder<
+               // Coefficient Storage API
+               coefficient_info<
+               // Symbol Storage API
+               deep_symbol_storage<
+               storage_bytes_used<
+               storage_block_info<
+               // Finite Field API
+               finite_field_math<typename fifi::default_field<Field>::type,
+               finite_field_info<Field,
+               // Factory API
+               final_coder_factory_pool<
+               // Final type
+               sparse_full_rlnc_encoder<Field
+                   > > > > > > > > > > > > > > > > >
+    { };
 
 }